requirements.txt
numpy==1.21.2
Pillow==8.3.1
tensorflow==2.5.1